// packages/core/types.d.ts
export interface VerifyResult {
    /** Verification status */
    status: 'ok' | 'failed' | 'error';
    /** Verification timestamp */
    timestamp: number;
}
// ...
export interface PaymentOptions {
    /** Telegram User ID */
    tg_use_id: string;
    /** Payment amount (in cents)
     * @see https://core.telegram.org/bots/api#labeledprice
     */
    amount: number;
    /** Label 
     * @see https://core.telegram.org/bots/api#labeledprice
     */
    label: string;
    /** Title */
    title: string;
    /** Description */
    description: string;
    /** Payload */
    payload?: string;
    /** Photo height */
    photo_height?: number;
    /** Photo width */
    photo_width?: number;
    /** Photo size */
    photo_size?: number;
    /** Photo URL */
    photo_url?: string;
    /** Callback URL */
    callback_url?: string;
}

/**
 * Payment Result
 */
export interface PaymentResult {
    /** Payment ID */
    tg_payment_id: number;
    /** Payment URL */
    tg_payment_url: string;
}
// ...
export interface InvoiceClosedResult {
    orderId: number;
    status: 'paid' | 'cancelled' | 'failed' | 'pending';
};

export interface OrderResult {
    /** Amount */
    amount: number;
    /** Application ID */
    application_id: string;
    /** Description */
    description: string;
    /** Label */
    label: string;
    /** Payload */
    payload: string;
    /** Status */
    status: number;
    /** Status description */
    status_description: string;
    /** Payment ID */
    tg_payment_id: string;
    /** Telegram User ID */
    tg_use_id: string;
    /** Title */
    title: string;
}
// ...
export interface PortexResponse<T> {
    ok: boolean;
    body: PortexResponseBody<T> | null;
    status: number;
    statusText: string;
    headers: Headers;
}

export interface PortexResponseBody<T> {
    ok: boolean;
    result: T | null;
    is_dev?: boolean;
    desc?: string;
}

export interface PortexRequestOptions {
    method?: 'GET' | 'POST' | 'PUT' | 'DELETE';
    data?: Record<string, any>;
    headers?: Record<string, string>;
}

/**
 * Request Interface
 */
export interface IPortex {
    /**
     * Call API
     */
    call<T>(path: string, options?: PortexRequestOptions): Promise<PortexResponse<T>>;
    /**
     * Initialize
     */
    init(): Promise<VerifyResult>;
    /**
     * Get start param
     */
    getStartParam(): string;
    /**
     * Check if user is verified
     */
    readonly isVerified: boolean;
    /**
     * Telegram Web App
     * @see https://core.telegram.org/bots/webapps
     */
    webApp: WebApp;
}